requirements.txt
numpy
opencv-python
ultralytics
pyserial
matplotlib
ipython

